Create readmidi
I've done almost no testing on this, but it worked two times in a row so far so hopefully its good. I'll iron things out tomorrow if there are problems, also I can make it return the loudness if need be, 100 for a note and 0 for a rest, but I choose not to include that. 

There is a solid chance the time value wont be correct if the song includes rests because of the way I wrote this. Keeping track of rests will fix that, but its not a big deal.